% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReViewCNV-package.R
\docType{package}
\name{ReViewCNV-package}
\alias{ReViewCNV-package}
\title{ReViewCNV: Visualization of germinal Copy Number Variants (CNVs)}
\description{
ReViewCNV is a ShinyApp for the visualization and clinical interpretation of germinal Copy Number Variants (CNVs). It visualizes CNVs from exome or gene panels sequencing identified by the algorithms EXCAVATOR2 and Mixer. It also accepts a tsv file with the identified CNVs as specified in README.
}
\author{
\strong{Maintainer}: Francesco Paparazzo \email{francescopaparazzo85@gmail.com}

Other contributors:
\itemize{
  \item CTGLab, IIT CNR [copyright holder, funder]
}

}
\keyword{internal}
