% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReViewCNV.R
\name{ReViewCNV}
\alias{ReViewCNV}
\title{ReViewCNV}
\usage{
ReViewCNV(host = "0.0.0.0", port = 3838)
}
\arguments{
\item{host}{The IPv4 address that the application should listen}

\item{port}{The TCP port that the application should listen on}
}
\value{
Runs a Shiny application in the default web browser.
}
\description{
ReViewCNV
}
\examples{
if (interactive()) {
  Plot_Visualization()
}
}
