% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReViewCNV.R
\name{ReViewCNV}
\alias{ReViewCNV}
\title{ReViewCNV}
\usage{
ReViewCNV(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shinyApp()}. Currently unused.}
}
\value{
Runs a Shiny application in the default web browser.
}
\description{
ReViewCNV
}
\examples{
if (interactive()) {
  Plot_Visualization()
}
}
